some revisions to the napkin calculations
From observations during the Thursday session as a pannelist I noticed that the bandwidth was higher than my original
estimate.  This might just be because I was a pannelist at the time.  I'll follow up today.  At the same time I
noticed that my estimate of 60% attendance might be high.  There were 120ish people on the call.

I was also thinking a bit more about the average distances travelled and the fact that they seemed too small.
Reflecting on the map that Julia showed off, most people were in EST and there were even some from the UK. Whitehorse
seemed too close to be the average.  There is an error in the calculation.  I tabulated the total people in the
fly and drive situation correctly but I forgot to account for the duplication of distances in calculating the total
distances used in the average.

Both these changes result in way higher costs. It also results in a way higher savings!